[WIP] Streaming Analytics Service v2 support, basic functions but not final.
Old direct uses of the Streaming Analytics Service REST API (mostly via
`RestUtils.java`, and use of responses) and the APIs added for Beam runner
in `com.ibm.streamsx.rest` have been changed to use factory methods to
create version-specific implementations.

Tested against the current SASv2 implementation in stage1, and current v1
services using Beam submit of `.sab` and remote build and submit, also with
collection of metrics via Streams REST API.

Not final because some required information about what will be available
in the credentials has not been finalized. These are marked with `FIXME`
in the code, but to summarize:

- Will IAM token URL be available in credentials, and by what name?
- Will API key be available in credentials, and by what name?
- What is the final name of the resources URL in credentials?
- What is the final name of the service ID in credentials?

If some are not available in credentials, we will have to figure out where
to put this information (eg. user manually adds to credentials, or elsewhere
in VCAP_SERVICES, or environment, etc.)

The above is required as a minimum, but some further improvements are
possible:

`RestUtils` and the `streamsx.rest` package had some similar code and error
handling that should be cleaned up and made common.

`RemoteStreamingAnalyticsServiceStreamsContext` is still aware of job ID
member name in responses, this should be pushed down into the version-specific
implementations, but this may require some changes to some of the contexts.

The construction of `StreamsConnection` implementations are not ideal because
2 of 3 start with URLs near the root, and common code finds the instances
from that, but v2 starts with the instance, constructs the root and forgets
the instance, which is then reconstructed by the common code. In particular
it would have problems if / when a subsequent version starts supporting more
than one instance per Streaming Analytics Service instance.

Authentication and SAS version are currently coupled, in practice and the
code, but conceptually authentication could be separated out into.

If SAS stabilizes on having a version-specific path prefix (current /v2,
next /v3, etc.) and the basic starting elements remain similar, the factory
might be able to load implementations by constructing the class name and
loading it, instead of switching on known versions.