[WIP] Finalize credentials content for Streaming Analytics Service V2.
Cloud team provided final member names. IAM URL will not be in credentials,
we will use the well-known Bluemix URL unless the host looks like a staging
server, then we construct one in the pattern of the staging endpoints.

Since this isn't quite as cheap as it was, cache the URL.

Also some cleanup of javadoc and constants.