Restore backward compatibility of StreamsConnection, StreamingAnalyticsConnection
The old class-based versions fo StreamingAnalyticsConnection and StreamsConnection have been restored by forwarding to the new interface-based versions.

Both classes are deprecated.

The two test classes have been reverted to the old APIs to check compatibility, with minor changes:

- additional info on some asserts
- more relaxed handling of cancel of non-existent job

The original API did not specify the latter; the method returns true or false but also declares it throws. The old test assumed it returns false, the new code will throw. I prefer the throw, but if reviewers disagree, the old behaviour can be restored.

Two tests are failing on master, before and after these changes, apparently because names of operators have changed. In particular, testOperators() and testProcessingElements() were expecting operators named IntegerPeriodicMultiSource and FunctionAnonymous_2 but are now seeing FunctionAnonymous and SupplierAnonymous, probably due to implementation changes in master.

These tests also depend on the order of operators returned from which influences the above and is probably not ideal, but it is not strictly an issue of order, some of the names do appear to have changed.