Remove StreamsRestFactory, move or inline methods.
The methods in StreamsRestFactory that were still in use were either inlined (twon instances in StreamsConnection) or split and moved to:

AbstractStreamingAnalyticsConnection.of() which calls one of:

StreamingAnalyticsConnectionV1.of() or
StreamingAnalyticsConnectionV2.of()

Util code moved to StreamsRestUtils.

Duplicate code from AbstractStreamingAnalyticsService and StreamingAnalyticsServiceV[12] was replaced with calls to one of the above.